Import loop fix
