+++ App/App.tsx
@@ -46,12 +46,3 @@ export default function App() {
         </AuthProvider>
 	);
 }
-
-export const styles = StyleSheet.create({
-	container: {
-		flex: 1,
-		backgroundColor: "#fff",
-		alignItems: "center",
-		justifyContent: "center",
-	},
-});
